Update CplusplusTest_cpp.cgt
Cleaned up the unused calls and updated the process. This reads a tuple coming in on port 0.  It then gets the value of the vals list and converts it to a double*.  It is then passed through an example function defined at the top of this file that returns 5.0 and prints it to the console.  Then it converts the double* back to a list and submits the tuple on port 0.